# baggage propagator clean up