--- a/tracer/src/Datadog.Trace/Propagators/W3CBaggagePropagator.cs
+++ b/tracer/src/Datadog.Trace/Propagators/W3CBaggagePropagator.cs
@@ -111,6 +111,13 @@ internal class W3CBaggagePropagator : IContextInjector, IContextExtractor
             return;
         }
 
+        if (!AnyCharRequiresEncoding(source, charsToEncode))
+        {
+            // no bytes require encoding, append the source string directly
+            sb.Append(source);
+            return;
+        }
+
         // this is an upper bound and will almost always be more bytes than we need
         var maxByteCount = Encoding.UTF8.GetMaxByteCount(source.Length);
         int byteCount;
@@ -124,7 +131,7 @@ internal class W3CBaggagePropagator : IContextInjector, IContextExtractor
 
             // slice the buffer down to the actual bytes written
             var stackBytes = stackBuffer[..byteCount];
-            EncodeBytesAndAppend(sb, source, stackBytes, charsToEncode);
+            EncodeBytesAndAppend(sb, stackBytes, charsToEncode);
             return;
         }
 #endif
@@ -138,7 +145,7 @@ internal class W3CBaggagePropagator : IContextInjector, IContextExtractor
 
             // slice the buffer down to the actual bytes written
             var bytes = buffer.AsSpan(0, byteCount);
-            EncodeBytesAndAppend(sb, source, bytes, charsToEncode);
+            EncodeBytesAndAppend(sb, bytes, charsToEncode);
         }
         finally
         {
@@ -146,18 +153,11 @@ internal class W3CBaggagePropagator : IContextInjector, IContextExtractor
         }
     }
 
-    private static void EncodeBytesAndAppend(StringBuilder sb, string source, Span<byte> bytes, HashSet<char> charsToEncode)
+    private static void EncodeBytesAndAppend(StringBuilder sb, Span<byte> bytes, HashSet<char> charsToEncode)
     {
-        if (!AnyByteRequiresEncoding(bytes, charsToEncode))
-        {
-            // no bytes require encoding
-            sb.Append(source);
-            return;
-        }
-
         foreach (var b in bytes)
         {
-            if (ByteRequiresEncoding(b, charsToEncode))
+            if (b < 0x20 || b > 0x7E || char.IsWhiteSpace((char)b) || charsToEncode.Contains((char)b))
             {
                 // encode byte as '%XX'
                 sb.Append($"%{b:X2}");
@@ -167,27 +167,19 @@ internal class W3CBaggagePropagator : IContextInjector, IContextExtractor
                 sb.Append((char)b);
             }
         }
+    }
 
-        return;
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        static bool ByteRequiresEncoding(byte b, HashSet<char> charsToEncode)
+    internal static bool AnyCharRequiresEncoding(string source, HashSet<char> charsToEncode)
+    {
+        foreach (var c in source)
         {
-            return b < 0x20 || b > 0x7E || char.IsWhiteSpace((char)b) || charsToEncode.Contains((char)b);
-        }
-
-        static bool AnyByteRequiresEncoding(Span<byte> bytes, HashSet<char> charsToEncode)
-        {
-            foreach (var b in bytes)
+            if (c < 0x20 || c > 0x7E || char.IsWhiteSpace(c) || charsToEncode.Contains(c))
             {
-                if (ByteRequiresEncoding(b, charsToEncode))
-                {
-                    return true;
-                }
+                return true;
             }
-
-            return false;
         }
+
+        return false;
     }
 
     internal static string Decode(string value)