## Summary of changes

Refactored how we generated the `baggage` header value in `W3CBaggagePropagator` to be faster and minimize heap allocations.

## Reason for change

Performance

## Implementation details

#### 1. `string` and `StringBuilder` management

We now use a single `StringBuilder` (from `StringBuilderCache`) to build the entire `baggage` header. Previously the code was trying to acquire a builder to encode keys and values while _already_ holding another builder for the header itself. With this change, we no longer allocate multiple `StringBuilder` instances that don't get reused, and we no longer allocation individual strings for:
- keys that need encoding
- values that need encoding
- every key/value pair (the string was built first to measure its length in UTF-8 bytes, but we don't need that anymore)

#### 2. checking earlier if string needs encoding

Before this PR, we always converted every key and value into UTF-8 bytes before checking if any bytes needed percent-encoding (aka url-encoding). When encoding was not needed (the common case), we could avoid the conversion and use the original string directly.

Now we check if the string's characters need encoding _before_ converting them to bytes. When encoding is not needed (the common case), we skip the conversion from string to UTF-8 bytes.

## Benchmarks

<details>
<summary>With allowed characters only (no percent-encoding required)</summary>

| Method | Runtime              | ItemCount | Mean        | Error    | StdDev   | Ratio | Gen0   | Allocated | Alloc Ratio |
|------- |--------------------- |---------- |------------:|---------:|---------:|------:|-------:|----------:|------------:|
| Before | .NET 6.0             | 1         |    91.63 ns | 0.309 ns | 0.289 ns |  1.00 | 0.0050 |      64 B |        1.00 |
| After  | .NET 6.0             | 1         |    49.84 ns | 0.162 ns | 0.152 ns |  0.54 | 0.0025 |      32 B |        0.50 |
|        |                      |           |             |          |          |       |        |           |             |
| Before | .NET Framework 4.6.1 | 1         |   253.30 ns | 1.867 ns | 1.655 ns |  1.00 | 0.0215 |     136 B |        1.00 |
| After  | .NET Framework 4.6.1 | 1         |   101.09 ns | 0.368 ns | 0.326 ns |  0.40 | 0.0063 |      40 B |        0.29 |
|        |                      |           |             |          |          |       |        |           |             |
| Before | .NET 6.0             | 5         |   343.57 ns | 1.953 ns | 1.731 ns |  1.00 | 0.0215 |     272 B |        1.00 |
| After  | .NET 6.0             | 5         |   139.36 ns | 2.696 ns | 3.104 ns |  0.41 | 0.0062 |      80 B |        0.29 |
|        |                      |           |             |          |          |       |        |           |             |
| Before | .NET Framework 4.6.1 | 5         | 1,303.19 ns | 6.138 ns | 5.742 ns |  1.00 | 0.1602 |    1019 B |        1.00 |
| After  | .NET Framework 4.6.1 | 5         |   328.44 ns | 0.956 ns | 0.848 ns |  0.25 | 0.0138 |      88 B |        0.09 |

</details>

<details>
<summary>With characters that require percent-encoding (whitespace and Unicode)</summary>

| Method | Runtime              | ItemCount | Mean       | Error    | StdDev   | Ratio | Gen0   | Allocated | Alloc Ratio |
|------- |--------------------- |---------- |-----------:|---------:|---------:|------:|-------:|----------:|------------:|
| Before | .NET 6.0             | 1         |   229.4 ns |  2.07 ns |  1.73 ns |  1.00 | 0.0815 |    1024 B |        1.00 |
| After  | .NET 6.0             | 1         |   105.5 ns |  0.76 ns |  0.71 ns |  0.46 | 0.0063 |      80 B |        0.08 |
|        |                      |           |            |          |          |       |        |           |             |
| Before | .NET Framework 4.6.1 | 1         |   990.2 ns |  2.04 ns |  1.60 ns |  1.00 | 0.4215 |    2656 B |        1.00 |
| After  | .NET Framework 4.6.1 | 1         |   279.9 ns |  1.46 ns |  1.37 ns |  0.28 | 0.0124 |      80 B |        0.03 |
|        |                      |           |            |          |          |       |        |           |             |
| Before | .NET 6.0             | 5         |   989.9 ns |  6.25 ns |  5.22 ns |  1.00 | 0.1469 |    1848 B |        1.00 |
| After  | .NET 6.0             | 5         |   438.5 ns |  4.18 ns |  3.91 ns |  0.44 | 0.0234 |     296 B |        0.16 |
|        |                      |           |            |          |          |       |        |           |             |
| Before | .NET Framework 4.6.1 | 5         | 4,989.7 ns | 50.79 ns | 45.02 ns |  1.00 | 1.6556 |   10447 B |        1.00 |
| After  | .NET Framework 4.6.1 | 5         | 1,243.1 ns |  7.70 ns |  7.20 ns |  0.25 | 0.0458 |     297 B |        0.03 |

</details>

## Test coverage

All `W3CBaggagePropagatorTests` are passing.

## Other details
Follow-up from #6158.
<!-- Fixes #{issue} -->

<!--  ⚠️ Note: where possible, please obtain 2 approvals prior to merging. Unless CODEOWNERS specifies otherwise, for external teams it is typically best to have one review from a team member, and one review from apm-dotnet. Trivial changes do not require 2 reviews. -->