@lucaspimentel 2024-11-13
### Merge activity

* **Nov 13, 5:31 PM EST**: A user merged this pull request with [Graphite](https://app.graphite.dev/github/pr/DataDog/dd-trace-dotnet/6266).